requirements.txt
APScheduler==3.11.0
apscheduler[redis]>=3.9 # APSchedulerとRedis連携用
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-APScheduler==1.13.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mailjet-rest>=1.3.0 # バージョンは適宜確認
oauthlib==3.2.2
packaging==24.2
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary # PostgreSQLアダプタを追加 (Render推奨)
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv # ローカル開発用
SQLAlchemy  # SQLAlchemyを追加
redis
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
rq
six==1.17.0
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
zoneinfo # Python 3.9未満の場合